got lib to load article json for 3 platforms
