--- app/babel.config.js
@@ -1,17 +1,0 @@
-module.exports = function (api) {
-  api.cache(true);
-  return {
-    presets: ['babel-preset-expo'],
-    plugins: [
-      'expo-router/babel',
-      [
-        'module-resolver',
-        {
-          alias: {
-            '@': '.',
-          },
-        },
-      ],
-    ],
-  };
-};
